Fix issue with inlined styles

+++ src/index.ts
@@ -20,7 +20,7 @@ import * as path from "path";
 import * as fs from "fs";
 
 class CrittersAstroPlugin extends Critters {
-	constructor(options: Options){
+	constructor(options: Options) {
 		super(options);
 	}
 
@@ -33,13 +33,15 @@ class CrittersAstroPlugin extends Critters {
 		const isStyleInlined = super.pruneSource(style, before, sheetInverse);
 		const name = style.$$name;
 
-		// @ts-ignore
-		const abs = path.join(this.options.path, name);
+		if (name) {
+			// @ts-ignore
+			const abs = path.join(this.options.path, name);
 
-		if (isStyleInlined) {
-			fs.rm(abs, () => {});
-		} else {
-			fs.writeFile(abs, sheetInverse, () => {});
+			if (isStyleInlined) {
+				fs.rm(abs, () => { });
+			} else {
+				fs.writeFile(abs, sheetInverse, () => { });
+			}
 		}
 
 		return isStyleInlined;
